+++ tests/test_login.py
@@ -1,13 +1,15 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-from shanbay import ShanbayException, AuthException, ServerException, Shanbay
+from shanbay import AuthException, ConnectException, Shanbay
 
 
 def test_login():
     s = Shanbay('root', 'root')
     s.login()
 
+
+def test_exception():
     try:
         Shanbay('rxxxoot', 'axxbcd').login()
     except AuthException:
@@ -19,5 +21,5 @@ def test_login():
             "https": "http://10.10.1.10:1080",
         }
         Shanbay('rooxxt', 'abcxx').login(proxies=proxies, timeout=3)
-    except ServerException:
+    except ConnectException:
         pass